# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SOWISO Etiquette API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user profile
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
          description: Optional user identifier; bearer token is the source of truth.
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProfile
      tags: [users]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "200":
          description: Profile created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: patchProfilePreferences
      tags: [users]
      summary: Patch user preferences (language and active region)
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: User identifier; must match the authenticated session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchProfilePreferencesBody"
      responses:
        "200":
          description: Profile preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProfile
      tags: [users]
      summary: Delete user profile
      responses:
        "200":
          description: Profile deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/profile/region:
    patch:
      operationId: updateActiveRegion
      tags: [users]
      summary: Update the user's active region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRegionBody"
      responses:
        "200":
          description: Region updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        email_verified:
          type: boolean
        birth_year:
          type: integer
          nullable: true
        gender_identity:
          type: string
          nullable: true
        gender_expression:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
          description: Computed alias for gender_identity
        age_group:
          type: string
          nullable: true
          description: Computed from birth_year (e.g. "18-25", "26-35")
        noble_score:
          type: integer
        ambition_level:
          type: string
          enum: [casual, professional, diplomatic]
        subscription_tier:
          type: string
          enum: [guest, traveller, ambassador]
        subscription_status:
          type: string
        language_code:
          type: string
        active_region:
          type: string
        region_history:
          type: array
          items:
            type: string
        is_admin:
          type: boolean
        suspended_at:
          type: string
          format: date-time
          nullable: true
        onboarding_completed:
          type: boolean
        country_of_origin:
          type: string
          nullable: true
        country_of_origin_locked_at:
          type: string
          format: date-time
          nullable: true
        situational_interests:
          type: array
          items:
            type: string
        objectives:
          type: array
          items:
            type: string
        interests_sports:
          type: array
          items:
            type: string
        interests_cuisine:
          type: array
          items:
            type: string
        interests_dress_code:
          type: array
          items:
            type: string
        payment_customer_id:
          type: string
          nullable: true
        trial_ends_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - email_verified
        - noble_score
        - ambition_level
        - subscription_tier
        - subscription_status
        - language_code
        - active_region
        - region_history
        - is_admin
        - onboarding_completed
        - objectives
        - interests_sports
        - interests_cuisine
        - interests_dress_code
        - created_at
    CreateProfileBody:
      type: object
      properties:
        birth_year:
          type: integer
          nullable: true
        gender_identity:
          type: string
          nullable: true
        gender_expression:
          type: string
          nullable: true
        ambition_level:
          type: string
          enum: [casual, professional, diplomatic]
        language_code:
          type: string
        active_region:
          type: string
      required:
        - id
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    UpdateProfileBody:
      type: object
      properties:
        birth_year:
          type: integer
          nullable: true
        gender_identity:
          type: string
          nullable: true
        gender_expression:
          type: string
          nullable: true
        ambition_level:
          type: string
          enum: [casual, professional, diplomatic]
        language_code:
          type: string
        active_region:
          type: string
        subscription_tier:
          type: string
          enum: [guest, traveller, ambassador]
    UpdateRegionBody:
      type: object
      properties:
        region_code:
          type: string
      required:
        - region_code
    PatchProfilePreferencesBody:
      type: object
      properties:
        language_code:
          type: string
          description: BCP-47 base language code (e.g. en, nl, fr).
        active_region:
          type: string
          description: ISO 3166-1 alpha-2 region code (e.g. GB, NL, US).
...